Census: quick breakout/refactor and documentation
